+++ src/main/java/nora/movlog/domain/Director.java
@@ -1,4 +1,4 @@
-package nora.movlog.entity;
+package nora.movlog.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -9,18 +9,16 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Data
 @Entity
-public class
-Genre {
+public class Director {
     /* 데이터 관리 및 조회용 */
-    @Id
-    @GeneratedValue
+    @Id @GeneratedValue
     private Long id;        // DB ID
 
 
-    /* 장르 데이터 */
-    private String name;    // 장르 이름
+    /* 감독 데이터 */
+    private String name;    // 한국 이름
 
-    public Genre(String name) {
+    public Director(String name) {
         this.name = name;
     }
 }
